requirements.txt
asgiref==3.7.2
asttokens==2.4.1
backcall==0.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.0
django-allauth==0.59.0
django-debug-toolbar==4.2.0
django-dump-load-utf8==0.0.4
django-environ==0.11.2
django-extensions==3.2.3
django-redis==5.4.0
django-shell-plus==1.1.7
djangorestframework==3.14.0
executing==2.0.0
flake8==6.1.0
idna==3.6
ipython==8.16.1
isort==5.12.0
jedi==0.19.1
matplotlib-inline==0.1.6
mccabe==0.7.0
oauthlib==3.2.2
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
Pygments==2.17.2
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
stripe==7.10.0
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.13
